requirements.txt
python-telegram-bot==20.7
python-binance==1.0.19
pandas>=2.0.0
numpy>=1.21.0
scikit-learn>=1.3.0
requests>=2.28.0
matplotlib>=3.5.0
python-dotenv>=1.0.0
ta>=0.10.0
joblib>=1.2.0
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
newsapi-python>=0.2.6